refactor: consolidating test utils
