+++ packages/api-graphql/__tests__/internals/server/generateClientWithAmplifyInstance.test.ts
@@ -1,9 +1,9 @@
-import * as raw from '../../../src';
 import { Amplify, AmplifyClassV6, ResourcesConfig } from '@aws-amplify/core';
 import { generateClientWithAmplifyInstance } from '../../../src/internals/server';
 import configFixture from '../../fixtures/modeled/amplifyconfiguration';
 import { Schema } from '../../fixtures/modeled/schema';
 import { V6ClientSSRRequest, V6ClientSSRCookies } from '../../../src/types';
+import { mockApiResponse, normalizePostGraphqlCalls } from '../../utils';
 
 const serverManagedFields = {
 	id: 'some-id',
@@ -24,66 +24,6 @@ const config: ResourcesConfig = {
 		},
 	},
 };
-
-/**
- *
- * @param value Value to be returned. Will be `awaited`, and can
- * therefore be a simple JSON value or a `Promise`.
- * @returns
- */
-function mockApiResponse(value: any) {
-	return jest
-		.spyOn((raw.GraphQLAPI as any)._api, 'post')
-		.mockImplementation(async () => {
-			const result = await value;
-			return {
-				body: {
-					json: () => result,
-				},
-			};
-		});
-}
-
-/**
- * For each call against the spy, assuming the spy is a `post()` spy,
- * replaces fields that are likely to change between calls (or library version revs)
- * with static values. When possible, on the unpredicable portions of these values
- * are replaced.
- *
- * ## THIS IS DESTRUCTIVE
- *
- * The original `spy.mocks.calls` will be updated *and* returned.
- *
- * For example,
- *
- * ```plain
- * headers.x-amz-user-agent: "aws-amplify/6.0.5 api/1 framework/0"
- * ```
- *
- * Is replaced with:
- *
- * ```plain
- * headers.x-amz-user-agent: "aws-amplify/latest api/latest framework/latest"
- * ```
- *
- * @param spy The Jest spy
- */
-function normalizePostGraphqlCalls(spy: jest.SpyInstance<any, any>) {
-	return spy.mock.calls.map((call: any) => {
-		// The 1st param in `call` is an instance of `AmplifyClassV6`
-		// The 2nd param in `call` is the actual `postOptions`
-		const [_, postOptions] = call;
-		const userAgent = postOptions?.options?.headers?.['x-amz-user-agent'];
-		if (userAgent) {
-			const staticUserAgent = userAgent.replace(/\/[\d.]+/g, '/latest');
-			postOptions.options.headers['x-amz-user-agent'] = staticUserAgent;
-		}
-		// Calling of `post` API with an instance of `AmplifyClassV6` has been
-		// unit tested in other test suites. To reduce the noise in the generated
-		// snapshot, we hide the details of the instance here.
-		return ['AmplifyClassV6', postOptions];
-	});
-}
 
 describe('server generateClient', () => {
 	beforeEach(() => {
